+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 See [changes since release][HEAD].
 
+## [3.3.0][] / 2018-03-17
+
 * Add "aria-live" to tablists for reporting tab changes to the user.
 * Add optional "aria-label" to tablists to improve screenreader reporting.
 * Give active tabpanels tabindex 0 to allow keyboard navigation to the active panel content.
@@ -154,4 +156,5 @@ to the README for current usage.
 [3.0.2]: https://github.com/IvyApp/ivy-tabs/compare/v3.0.1...v3.0.2
 [3.1.0]: https://github.com/IvyApp/ivy-tabs/compare/v3.0.2...v3.1.0
 [3.2.0]: https://github.com/IvyApp/ivy-tabs/compare/v3.1.0...v3.2.0
-[HEAD]: https://github.com/IvyApp/ivy-tabs/compare/v3.2.0...master
+[3.3.0]: https://github.com/IvyApp/ivy-tabs/compare/v3.2.0...v3.3.0
+[HEAD]: https://github.com/IvyApp/ivy-tabs/compare/v3.3.0...master